Bump to jQuery 1.11.2
This is mainly because something strange seems to have happened to the
existing rails-assets packaging of jquery 1.11.1. There's no longer a
jquery/jquery.js file, but instead only a jquery-legacy/jquery.js file
in the 1.11.1 gem. This was breaking fresh new builds, if there wasn't a
cached version of the old gem around. Let's try to updating to this
newer package which appears fixed.