Make all API urls relative for easier staging testing.
This is so links on our staging site will resolve to the staging server,
rather than redirecting off to production. I suppose this might make it
less convenient for local development, since these APIs won't be running
locally, so we may want to revisit more globally to make this
configurable.