Update to Swagger-UI v3. Update all Swagger docs to v2 of specification.
In order to use v3 of the UI, we need to upgrade some of our swagger
specification docs from v1.2 to 2.0. This was done through the
"swagger-tools convert" command, which programatically did this.
Although, we had some validation issues in our existing files, so those
first had to be fixed (so I'm also committing the fixed v1.2 documents
that were actually used for conversion).

Also switched all the v2 docs to use YAML instead of JSON to hopefully
make these a bit easier to maintain.