Add documentation for the errors array.
Since declaring this as an array, we now need to define the structure
inside.

This also fixes the rake task that validates the swagger specs to run
against the "source" files as well as the "build" files. This makes it
easier to run this task locally and validate what's being worked on in
development.