Switch from middleman-sprockets to external pipeline & webpacker.
This external pipeline is now the preferred approach to handling
assets in webpacker.

Also update all the JavaScript dependencies to the latest versions.

Fighting with gem updates related to sass, sassc, and
middleman-sprockets is what led me down this path, but since
middleman-sprockets is sort of deprecated in favor of the external
pipeline approach, hopefully this will be better supported. This also
gets us on newer javascript dependencies, which should be easier to
update via NPM/Yarn.