Begin to setup CI using GitHub Actions.
So that this repo doesn't depend on our internal Jenkins server (which
we started to use due to backlogs in Circle builds in the NREL org).
This should more clearly provide test results and build results for this
public repo. This also cleans up vestiges of the Circle CI setup that
was probably confusing matters, since it wasn't really being used.